fix: only show buy with Stripe option if URL is available
